// src/app/models/Utilisateur.ts
export interface Utilisateur {
    nomUtilisateur: string;
    motDePasse: string;
    codeResponsable: string;
    codeDroitAcces: string;
    nomPrenom: string;
    tel: string;
    actif: boolean;
    creerPar: string;
    dateCreation: Date;
    dateDerniereModification: Date;
    derniereModificationPar: string;
  }
  